+++ src/main/java/be/baur/sds/SDS.java
@@ -1,5 +1,10 @@
 package be.baur.sds;
 
+import java.io.IOException;
+import java.io.Reader;
+
+import be.baur.sda.serialization.SDAParseException;
+import be.baur.sds.serialization.SDSParseException;
 import be.baur.sds.serialization.SDSParser;
 import be.baur.sds.validation.SDAValidator;
 import be.baur.sds.validation.Validator;
@@ -11,16 +16,22 @@ public final class SDS {
 	
 	private SDS() {} // cannot construct this
 
-	
+	private static SDSParser PARSER = new SDSParser(); // singleton parser
+
 	/**
-	 * Returns a new instance of the default SDS parser.
+	 * Creates a schema from a character input stream in SDS format, using the
+	 * default SDS parser.
 	 * 
-	 * @return an {@link SDSParser}
+	 * @return a schema
+	 * @throws IOException       if an I/O operation failed
+	 * @throws SDAParseException if an SDA parse exception occurs
+	 * @throws SDSParseException if an SDS parse exception occurs
+	 * @see SDSParser
 	 */
-	public static SDSParser parser() {
-		return new SDSParser();
+	public static Schema parse(Reader input) throws IOException, SDAParseException, SDSParseException {
+		return PARSER.parse(input);
 	}
-	
+
 	
 	/**
 	 * Returns a new instance of the default SDS validator.