+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ This is (mostly) a compatibility release for sda-core 2.2.x.
 
 - `Removed` Interface be.baur.sds.serialization.Parser.
 - `Changed` NodeType.getName to NodeType.getTypeName.
+- `Changed` SDS.parser() to SDS.parse().
 - `Renamed` SchemaException to SDSParseException.
 - `Renamed` MixedType to DataType.